// frontend/src/api/generated/models/sharedTicket.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable ticket tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */
import type { SharedAttachment } from './sharedAttachment';
import type { SharedComment } from './sharedComment';
import type { SharedStatus } from './sharedStatus';
import type { TicketType } from './ticketType';

export interface SharedTicket {
  identifier: string;
  title: string;
  type: TicketType;
  status: SharedStatus;
  assignee?: string | null;
  estimate?: number | null;
  minutes_logged?: number | null;
  comments?: SharedComment[] | null;
  attachments?: SharedAttachment[] | null;
}